WIP: Fixes

+++ REDsym/actions.py
@@ -22,19 +22,19 @@ def update_wm2():
     print("There are:", len(deleted_dir_folders), "symlinks to delete")
 
     for deleted_dir_folder in deleted_dir_folders:
-        if rootdir_wcd in deleted_dir_folder:
+        if rootdir_wcd and rootdir_wcd in deleted_dir_folder:
             symlink = db.get_symlink_WCD_from_dir(deleted_dir_folder)
             if symlink:
                 db.delete_symlink_WCD(symlink)
                 os.remove(symlink)
-        if rootdir_red in deleted_dir_folder:
+        if rootdir_red and rootdir_red in deleted_dir_folder:
             symlink = db.get_symlink_RED_from_dir(deleted_dir_folder)
             if symlink:
                 db.delete_symlink_RED(symlink)
                 os.remove(symlink)
 
     for new_dir_folder in new_dir_folders_only_music:
-        if rootdir_wcd in new_dir_folder:
+        if rootdir_wcd and rootdir_wcd in new_dir_folder:
             audio_info_wm2 = db.get_meta_from_dir_WCD(new_dir_folder)
             artist_name, album_name = REDsym.util.audio_dir_filename_wcd_wm2(new_dir_folder)
             artist_name_valid = REDsym.util.filenamefy(artist_name)
@@ -52,13 +52,14 @@ def update_wm2():
                 db.insert_symlink_WCD(new_dir_folder, target)
             except FileExistsError:
                 print("symlink exists '%s' -> '%s'" % (fix_text(source), target))
-        if rootdir_red in new_dir_folder:
+        if rootdir_red and rootdir_red in new_dir_folder:
             audio_info_wm2 = db.get_meta_from_dir_RED(new_dir_folder)
             artist_name, album_name = REDsym.util.audio_dir_filename_red_wm2(new_dir_folder)
             artist_name_valid = REDsym.util.filenamefy(artist_name)
             album_name_valid = REDsym.util.filenamefy(album_name)
-            if not os.path.exists(redsym_dir+artist_name_valid):
-                os.makedirs(redsym_dir+ artist_name_valid)
+            target_dir = os.path.join(redsym_dir, artist_name_valid)
+            if not os.path.exists(target_dir):
+                os.makedirs(target_dir)
             new_album_path_folder = REDsym.util.get_music_dir(new_dir_folder)
             source = os.path.join(new_dir_folder, new_album_path_folder)
             target = os.path.join(redsym_dir, artist_name_valid, album_name_valid)
@@ -66,6 +67,8 @@ def update_wm2():
             if os.path.exists(target):
                 target = target + ' ID: ' + str(audio_info_wm2['TorrentId'])
             try:
+                print('Source:', source)
+                print('Target:', target)
                 os.symlink(source, target)
                 db.insert_symlink_RED(new_dir_folder, target)
             except FileExistsError:
